Add inflation driver and add deswell to regular driver
This commit also includes a little spring cleaning and a
restructure of damped_solve() such that it can start with
a steady solve if this is recommended. It now returns a
suggestion as to whether a steady solve would be a good
idea or not as well as a suggested timestep.